// apps/backend/src/common/constants/enums.ts
export enum UserRole {
  USER = 'USER',
  MODERATOR = 'MODERATOR',
  ADMIN = 'ADMIN',
}

export enum NotificationType {
  SYSTEM_ANNOUNCEMENT = 'SYSTEM_ANNOUNCEMENT',
  GENERAL = 'GENERAL',
}

export enum SubscriptionStatus {
  ACTIVE = 'ACTIVE',
  EXPIRED = 'EXPIRED',
  CANCELED = 'CANCELED',
  GRACE = 'GRACE',
}

export enum PlanType {
  WEEKLY = 'WEEKLY',
  MONTHLY = 'MONTHLY',
  YEARLY = 'YEARLY',
  LIFETIME = 'LIFETIME',
}

export enum Platform {
  ANDROID = 'ANDROID',
  IOS = 'IOS',
}

export enum PurchaseStatus {
  PENDING = 'PENDING',
  PURCHASED = 'PURCHASED',
  CANCELED = 'CANCELED',
  EXPIRED = 'EXPIRED',
  REVOKED = 'REVOKED',
}

export enum Language {
  EN = 'EN',
  VI = 'VI',
}

export enum Unit {
  MM = 'MM',
  INCH = 'INCH',
}